pdf2anki/requirements.txt
PyMuPDF==1.26.6
Pillow==10.4.0
pytesseract==0.3.10
openai>=2.0.0
python-dotenv>=1.0.0
